% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsANOVA.R
\docType{class}
\name{lsANOVA-class}
\alias{lsANOVA-class}
\title{lsANOVA}
\arguments{
\item{x}{(data.frame) The dataset to be processed.}

\item{factors}{(character) The columns to be trated be factors.}

\item{marks}{(character) The columns to be trated be marks.}

\item{max_p}{(numeric, default: 0.05) The max p.value of F test.}

\item{min_bartlett}{(numeric, default: 0.1) The min p.value of bartlett test.}
}
\value{
(lsANOVA) An object include the acceptable F test.
}
\description{
Class to describe available F test result.
}
\section{Methods}{

\describe{
\item{\code{initialize(x = NA_real_, factors = NA_real_, marks = NA_real_,
  max_p = 0.05, min_bartlett = 0.1)}}{Calculate the acceptable F test result.}
}}
\author{
Losses Don
}
\keyword{ANOVA}

