% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsANOVA.R
\name{lsANOVA}
\alias{lsANOVA}
\title{lsANOVA}
\usage{
lsANOVA(...)
}
\arguments{
\item{...}{params to pass to lsANOVA object.}
}
\value{
(lsANOVA) An object include the acceptable F test.
}
\description{
Generate a lsANOVA Object.
}
\note{
See lsANOVA-class.
}
\author{
Losses Don
}
\keyword{ANOVA}

