% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsANOVA.R
\name{lsANOVA.read}
\alias{lsANOVA.read}
\title{lsANOVA.read}
\usage{
lsANOVA.read(...)
}
\arguments{
\item{...}{params to pass to lsANOVARead object.}
}
\value{
(lsANOVARead) An object include the description of acceptable F test.
}
\description{
Generate a lsANOVARead Object.
}
\note{
See lsANOVARead-class.
}
\author{
Losses Don
}
\keyword{ANOVA}

