% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsANOVA.R
\docType{class}
\name{lsANOVARead-class}
\alias{lsANOVARead-class}
\title{lsANOVARead}
\arguments{
\item{x}{(lsANOVA) The object generated by lsANOVA.}
}
\value{
(lsANOVARead) An object include the description of acceptable F test.
}
\description{
Get the description of the ANOVA result.
}
\section{Methods}{

\describe{
\item{\code{initialize(x = NA_real_)}}{generate the detail information of each group devided by acceptable ANOVA result.}
}}
\note{
While using print function, you can add a param 'round = F' to show the full result of description (holy long).
}
\author{
Losses Don
}
\keyword{ANOVA}

